Add files via upload
Example of the CSV file that will be going into the database. I still need to clean it up. I will be removing cooking time and difficulty out of laziness. As for cooking instructions, it was causing to many issues so it was also removed. For cleaning up, the formatting on Allergens and the nutrition facts are weird so I will fix it later today. There is a csv file for all letters of the alphabet totaling 25 because x had none. I will upload a zip of them all when they are cleaned.